Update live tests
The live tests just print out information and aren't really doing
"tests" in the sense that they don't assert anything and don't fail.

This is because not enough API functionality has been implemented
to actually perform validation to check if the tests are successful.

In the future, the plan is to extend them to actually do a series of
tasks (like create a phonebook, bulk upload contacts, upload a single
contact, etc) and then validate they were done successfully by
downloading the contents of the phonebook.