Modify unit test requests to use requests_mock, add __repr__ for objects
In order to add __repr__ to objects, we need to get the admin username
in the __init__ method, which we can do by a call to /agents/ and then
looking for the "owner" of the first agent. However, this seemingly minor
change breaks the entire test infrastructure, because MagicMock can
only be applied to a session AFTER it has been created, and if a web
request is made in __init__ it cannot be mocked.

We can solve this by using requests_mock to mock web requests!